+++ src/main/java/com/softlab/hospital/web/api/SuperManagerApi.java
@@ -1,8 +1,10 @@
 package com.softlab.hospital.web.api;
 
+import com.softlab.hospital.common.ErrorMessage;
 import com.softlab.hospital.common.HosExection;
 import com.softlab.hospital.common.RestData;
 import com.softlab.hospital.common.util.JsonUtil;
+import com.softlab.hospital.common.util.VerifyUtil;
 import com.softlab.hospital.core.model.User;
 import com.softlab.hospital.core.model.vo.DoctorVo;
 import com.softlab.hospital.service.ManagerService;
@@ -39,6 +41,10 @@ public class SuperManagerApi {
     public RestData selectDoctorByCondition(@RequestBody DoctorVo doctorVo, HttpServletRequest request) {
         logger.info("GET selectDoctorByCondition : " + JsonUtil.getJsonString(doctorVo));
 
+        if (1 != VerifyUtil.verifyType(request)){
+            return new RestData(1, ErrorMessage.OPERATIOND_ENIED);
+        }
+
         try {
             return superManagerService.selectDoctorByCondition(doctorVo);
         } catch (HosExection e) {
@@ -49,6 +55,10 @@ public class SuperManagerApi {
     @RequestMapping(value = "/manager", method = RequestMethod.POST)
     public RestData postAddManager(@RequestBody User user, HttpServletRequest request) {
         logger.info("POST postAddManager : " + JsonUtil.getJsonString(user));
+
+        if (1 != VerifyUtil.verifyType(request)){
+            return new RestData(1, ErrorMessage.OPERATIOND_ENIED);
+        }
 
         try {
             return superManagerService.postAddManager(user);
@@ -61,6 +71,10 @@ public class SuperManagerApi {
     public RestData deleteManagerById(@PathVariable Integer systemId, HttpServletRequest request) {
         logger.info("DELETE deleteManager : " + systemId);
 
+        if (1 != VerifyUtil.verifyType(request)){
+            return new RestData(1, ErrorMessage.OPERATIOND_ENIED);
+        }
+
         try{
             return superManagerService.deleteManager(systemId);
         } catch (HosExection e){
@@ -72,6 +86,10 @@ public class SuperManagerApi {
     public RestData updateManager(@RequestBody User user, HttpServletRequest request) {
         logger.info("DELETE deleteManager : " + JsonUtil.getJsonString(user));
 
+        if (1 != VerifyUtil.verifyType(request)){
+            return new RestData(1, ErrorMessage.OPERATIOND_ENIED);
+        }
+
         try{
             return superManagerService.updateManager(user);
         } catch (HosExection e){
@@ -80,11 +98,30 @@ public class SuperManagerApi {
     }
 
     @RequestMapping(value = "/manager-all", method = RequestMethod.GET)
-    public RestData selectAllManager() {
+    public RestData selectAllManager(HttpServletRequest request) {
         logger.info("GET selectAllManager : ");
+
+        if (1 != VerifyUtil.verifyType(request)){
+            return new RestData(1, ErrorMessage.OPERATIOND_ENIED);
+        }
 
         try{
             return superManagerService.selectAllManager();
+        } catch (HosExection e){
+            return new RestData(1, e.getMessage());
+        }
+    }
+
+    @RequestMapping(value = "/manager-part", method = RequestMethod.POST)
+    public RestData selectAllManager(@RequestBody User user, HttpServletRequest request) {
+        logger.info("GET selectPartManager : ");
+
+        if (1 != VerifyUtil.verifyType(request)){
+            return new RestData(1, ErrorMessage.OPERATIOND_ENIED);
+        }
+
+        try{
+            return superManagerService.selectPartManager(user);
         } catch (HosExection e){
             return new RestData(1, e.getMessage());
         }
